// src/app/Model/TenpOrderSummaryData.ts
export interface TempOrderSummaryData {
  pCustomerCode: string
  pProjectCode: string
  pSelectedCount: number
  pSelectedSE: any[]
  pSelectedProd: any[]
  pSelectedWT: any[]
  pSelectedQty: any[]
  pSelectedPostID: any[]
  pSelectedScheduled: any[]
  pSelectedWBS1: any[]
  pSelectedWBS2: any[]
  pSelectedWBS3: any[]
  pWBS1: string
  pWBS2: string
  pWBS3: string
  pOrderNo: any[]
  StructProd: any[]
}